+++ src/config-apollo/src/ClientFactory.php
@@ -9,6 +9,7 @@ declare(strict_types=1);
  * @contact  group@hyperf.io
  * @license  https://github.com/hyperf/hyperf/blob/master/LICENSE
  */
+
 namespace Hyperf\ConfigApollo;
 
 use Hyperf\Contract\ConfigInterface;
@@ -32,13 +33,14 @@ class ClientFactory
         $callbacks = [];
         foreach ($namespaces as $namespace => $callable) {
             // If does not exist a user-defined callback, then delegate to the dafault callback.
-            if (! is_numeric($namespace) && is_callable($callable)) {
+            if (!is_numeric($namespace) && is_callable($callable)) {
                 $callbacks[$namespace] = $callable;
             }
         }
         $httpClientFactory = function (array $options = []) use ($container) {
             return $container->get(GuzzleClientFactory::class)->create($options);
         };
-        return make(Client::class, compact('option', 'callbacks', 'httpClientFactory', 'config'));
+
+        return make($config->get('apollo.fetch_client', Client::class), compact('option', 'callbacks', 'httpClientFactory', 'config'));
     }
 }
