+++ src/config-apollo/src/Process/ConfigFetcherProcess.php
@@ -88,16 +88,13 @@ class ConfigFetcherProcess extends AbstractProcess
                 }
             }
         };
-        while (true) {
-            $callbacks = [];
-            $namespaces = $this->config->get('apollo.namespaces', []);
-            foreach ($namespaces as $namespace) {
-                if (is_string($namespace)) {
-                    $callbacks[$namespace] = $ipcCallback;
-                }
+        $callbacks = [];
+        $namespaces = $this->config->get('apollo.namespaces', []);
+        foreach ($namespaces as $namespace) {
+            if (is_string($namespace)) {
+                $callbacks[$namespace] = $ipcCallback;
             }
-            $this->client->pull($namespaces, $callbacks);
-            sleep($this->config->get('apollo.interval', 5));
         }
+        $this->client->fetch($namespaces, $callbacks);
     }
 }
